core/requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.4.0
astroid==3.2.3
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
decorator==4.4.2
dill==0.3.8
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.35.1
imageio-ffmpeg==0.5.1
isort==5.13.2
lxml==5.3.0
magic-filter==1.0.12
mccabe==0.7.0
moviepy==1.0.3
multidict==6.0.4
numpy==2.1.1
pillow==10.4.0
platformdirs==4.2.2
proglog==0.1.10
pydantic==2.5.3
pydantic_core==2.14.6
pylint==3.2.5
python-dotenv==1.0.0
python-pptx==1.0.2
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
tomlkit==0.13.0
tqdm==4.66.5
typing_extensions==4.9.0
urllib3==2.2.3
XlsxWriter==3.2.0
yadisk==3.1.0
yarl==1.9.4
